Move C++ `TableType` to its own file
With some small tests
